% Auto-generated: do not edit by hand
\name{fullCalendarComponent}

\alias{fullCalendarComponent}

\title{FullCalendarComponent component}

\description{

}

\usage{
fullCalendarComponent(id=NULL, clickedEvent=NULL, dateClicked=NULL, editable=NULL,
events=NULL, initialDate=NULL, selectable=NULL)
}

\arguments{
\item{id}{Character. }

\item{clickedEvent}{Named list. }

\item{dateClicked}{Character. }

\item{editable}{Logical. }

\item{events}{List of lists containing elements 'title', 'start', 'end', 'classname', 'context'.
those elements have the following types:
  - title (character; optional)
  - start (character; optional)
  - end (character; optional)
  - classname (character; optional)
  - context (character; optional)s. }

\item{initialDate}{Character. }

\item{selectable}{Logical. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

